Replace Unicode with ASCII in a few places
The JSON backend extracts function code verbatim into JSON format.
Unfortunately, the Unicode "Less-than Or Equal To" character, when embedded in
JSON, even when escaped using OCaml `String.escaped` does not successfully parse
with Python (3.10.12).

For now, at least, revert the Unicode symbol to the ASCII equivalent, which is
exactly what is used in the corresponding code and thus might even reduce the
chance of confusion.